iNES header now its own module
